requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.0
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.9.0
tiktoken==0.5.2
httpx==0.25.2
redis==5.0.1
prometheus-client==0.19.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
psutil==5.9.8
numpy<2.0
huggingface-hub<0.21.0